requirements.txt
fastapi
uvicorn[standard]
google-generativeai
psycopg2-binary
qdrant-client
pgvector
pydantic
python-dotenv
pydantic-settings
asyncpg
supabase
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx
